pyproject.toml
[tool.poetry]
name = "pyjobs_challenge_one"
version = "0.1.0"
description = ""
authors = ["ThiagoDiasV <thiago76ers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-filter = "^2.3.0"
pytest-django = "^3.9.0"
codecov = "^2.1.8"
python-decouple = "^3.3"
psycopg2-binary = "^2.8.5"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]
django = "^3.0.8"
djangorestframework = "^3.11.0"
pytest = "^6.0.1"
pytest-cov = "^2.10.0"
pylama = "^7.7.1"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
